chore: control LED with tilt
~ reverted back to int for temp- converting temp on server allows for
  conversions in other units as well, and not just celcius, and also
  allows to retrieve more sensitive values
~ changed the sequence of transmitting json

+++ include/Config.h
@@ -45,7 +45,7 @@ typedef struct {
         int16_t X, Y, Z;
     } Gyro;
 
-    double Temperature;
+    int16_t Temperature;
 } MotionValues;
 
 typedef struct {
